+++ Raytracer/MetalMaterial.swift
@@ -10,15 +10,15 @@ import Foundation
 import simd
 
 public class MetalMaterial: Material {
-	public var albedo: float3
-	public var fuzziness: Float
+	public var albedo: double3
+	public var fuzziness: Double
 	
-	public init(albedo a: float3, fuzziness f: Float) {
+	public init(albedo a: double3, fuzziness f: Double) {
 		albedo = a
 		fuzziness = f
 	}
 	
-	public override func scatter(rayIn: Ray, rec: HitRecord) -> (attenuation: float3, scattered: Ray)? {
+	public override func scatter(rayIn: Ray, rec: HitRecord) -> (attenuation: double3, scattered: Ray)? {
 		let reflected = reflect(normalize(rayIn.direction), n: rec.normal)
 		let scattered = Ray(origin: rec.point, direction: reflected + fuzziness * randomInUnitSphere(), time: rayIn.time)
 		let attenuation = albedo
